apps/user/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url','username', 'first_name', 'last_name',
                  'last_login', 'email', 'is_staff')
